refactor: 화살표 이미지 common 폴더로 분리
